Crate framenode_runtime

Re-exports

pub use assets;

pub use eth_bridge;

pub use frame_system;

pub use multicollateral_bonding_curve_pool;

pub use xst;

pub use oracle_types::*;

Modules

api

constants

Constant values used within the runtime.

migrations

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

oracle_types

Types used by oracle related pallets

Macros

balance

construct_runtime

Construct a runtime, with the given name and the given pallets.

fixed

parameter_types

Create new implementations of the Get trait.

Structs

AssetName

AssetSymbol

BagThresholds

BalanceWrapper

Wrapper type which extends Balance serialization, used for json in RPC’s.

BasicDeposit

BetterUnsignedThreshold

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1 and adding 0.

BlockHashCount

BlockLength

BlockWeights

Block weights base values and limits.

BondingDuration

CeresAssetId

CeresPerDay

ConstantFeeMultiplier

ContentSource

CouncilCollectiveMaxMembers

CouncilCollectiveMaxProposals

CouncilCollectiveMotionDuration

CreationFee

CrowdloanReward

A vested reward for crowdloan.

DEXIdValue

DemeterAssetId

DemocracyCooloffPeriod

DemocracyEnactmentPeriod

DemocracyFastTrackVotingPeriod

DemocracyInstantAllowed

DemocracyLaunchPeriod

DemocracyMaxProposals

DemocracyMaxVotes

DemocracyMinimumDeposit

DemocracyPreimageByteDeposit

DemocracyVotingPeriod

DepositBase

DepositFactor

DisabledValidatorsThreshold

ElectionBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionLookahead

ElectionsCandidacyBond

ElectionsDesiredMembers

13 members initially, to be increased to 23 eventually.

ElectionsDesiredRunnersUp

ElectionsModuleId

ElectionsTermDuration

ElectionsVotingBondBase

ElectionsVotingBondFactor

EpochDuration

EthNetworkId

ExistentialDeposit

ExistentialDeposits

ExpectedBlockTime

FarmingRewardDoublingAssets

FieldDeposit

GenesisConfig

GetADARAccountId

GetBaseAssetId

GetBurnUpdateFrequency

GetBuyBackAccountId

GetBuyBackAssetId

GetBuyBackDexId

GetBuyBackPercentage

GetBuyBackSupplyAssets

GetCrowdloanRewardsAccountId

GetCrowdloanRewardsTechAccountId

GetDaiAssetId

GetDefaultSubscriptionFrequency

GetDotAssetId

GetEthAssetId

GetFarmingRewardsAccountId

GetFarmingRewardsTechAccountId

GetFee

GetKsmAssetId

GetLiquidityProxyAccountId

GetLiquidityProxyTechAccountId

GetMarketMakerRewardsAccountId

GetMarketMakerRewardsTechAccountId

GetMbcPoolFreeReservesAccountId

GetMbcPoolFreeReservesTechAccountId

GetMbcPoolRewardsAccountId

GetMbcPoolRewardsTechAccountId

GetMbcReservesAccountId

GetMbcReservesTechAccountId

GetNumSamples

GetParliamentAccountId

GetPswapAssetId

GetPswapDistributionAccountId

GetPswapDistributionTechAccountId

GetSyntheticBaseAssetId

GetTotalBalance

GetUsdAssetId

GetValAssetId

GetXSTPoolPermissionedAccountId

GetXSTPoolPermissionedTechAccountId

GetXorAssetId

GetXorFeeAccountId

GetXorFeeTechAccountId

GetXstAssetId

GetXstPoolConversionAssetId

HermesAssetId

ImOnlineUnsignedPriority

LiquiditySourceFilter

Generalized filtration mechanism for listing liquidity sources.

LiquiditySourceId

Identification of liquidity source.

MaxAdditionalFields

MaxAuthorities

MaxElectableTargets

… and all of the validators as electable targets. Whilst this is the case, we cannot and shall not increase the size of the validator intentions.

MaxElectingVoters

We take the top 12500 nominators as electing voters..

MaxIterations

MaxKeys

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxRegistrars

MaxSignatories

MaxSubAccounts

MaximumCeresInStakingPool

MinSolutionScoreBump

MinimumPeriod

NoPreimagePostponement

NposCompactSolution24

NposSolutionPriority

OffchainRepeat

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

OffencesWeightSoftLimit

OffendingValidatorsThreshold

OnChainSeqPhragmen

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PresetWeightInfo

ReferralsReservesAcc

ReferrerWeight

RemovePendingOutgoingRequestsAfter

RemoveTemporaryPeerAccountIds

ReportLongevity

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeOnPswapBurnedAggregator

SS58Prefix

SchedulerMaxWeight

SessionDuration

SessionOffset

SessionPeriod

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxRefunds

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

SoraParliamentShare

StakingBenchmarkingConfig

A reasonable benchmarking config for staking pallet.

SubAccountDeposit

SwapOutcome

Amount of output tokens from either price request or actual exchange.

TechnicalCollectiveMaxMembers

TechnicalCollectiveMaxProposals

TechnicalCollectiveMotionDuration

TrackPendingIncomingRequestsAfter

TransactionByteFee

TransferFee

U128CurrencyToVote

An implementation of CurrencyToVote tailored for chain’s that have a balance type of u128.

UncleGenerations

UnsignedPhase

ValBurnedAggregator

ValRewardCurve

Version

WithdrawFee

XorBurnedWeight

XorIntoValBurnedWeight

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

DispatchClass

A generalized group of dispatch types.

Event

FilterMode

LiquiditySourceType

Enumaration of all available liquidity sources.

OriginCaller

StakerStatus

Indicates the initial status of the staker.

SwapAmount

Used to identify intention of caller either to transfer tokens based on exact input amount or exact output amount.

SwapVariant

Used to identify intention of caller to indicate desired input amount or desired output amount. Similar to SwapAmount, does not hold value in order to be used in external API.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

FARMING_PSWAP_PER_DAY

FARMING_REFRESH_FREQUENCY

FARMING_VESTING_COEFF

FARMING_VESTING_FREQUENCY

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

FromGenericPair

Abstract trait to get data type from generic pair name and data.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

LiquiditySource

Indicates that particular object can be used to perform exchanges.

MultiplierUpdate

Something that can convert the current multiplier to the next one.

OnPswapBurned

General trait for passing pswap amount burned information to required pallets.

OnUnbalanced

Handler for when some currency “account” decreased in balance for some reason.

OnValBurned

General trait for passing on the amount of burned VAL.

Randomness

A trait that is able to provide randomness.

ScheduleNamed

A type that can be used as a scheduler.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

debug

Print out the debuggable type.

fixed_from_basis_points

Create fraction as Fixed from BasisPoints value.

native_version

The version infromation used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Amount

AssetId

Assets

AssetsConfig

Authorship

Babe

BabeConfig

BagsList

Balance

BalancePrecision

Balances

BalancesConfig

Band

BasisPoints

Type definition representing financial basis points (1bp is 0.01%)

Beefy

BeefyConfig

BeefyId

Identity of a BEEFY authority using ECDSA as its crypto.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

BridgeMultisig

BridgeMultisigConfig

CeresGovernancePlatform

CeresLaunchpad

CeresLiquidityLocker

CeresStaking

CeresTokenLocker

CheckedExtrinsic

Extrinsic type that has already been checked.

Council

CouncilConfig

Currencies

DEXAPI

DEXAPIConfig

DEXId

Identification of DEX.

DEXManager

DEXManagerConfig

DemeterFarmingPlatform

Democracy

DemocracyConfig

DigestItem

Digest item type.

ElectionProviderMultiPhase

ElectionsPhragmen

ElectionsPhragmenConfig

EthBridge

EthBridgeConfig

Executive

Executive: handles dispatch to the various modules.

Farming

Fixed

Basic type representing assets quantity.

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

HermesGovernancePlatform

Historical

Identity

ImOnline

ImOnlineConfig

ImOnlineId

An i’m online identifier using sr25519 as its crypto.

Index

Index of a transaction in the chain.

IrohaMigration

IrohaMigrationConfig

LiquidityProxy

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Moment

MulticollateralBondingCurvePool

MulticollateralBondingCurvePoolConfig

Multiplier

Fee multiplier.

Multisig

NetworkId

Offences

OnChainAccuracy

The accuracy type used for genesis election provider;

OracleProxy

PeriodicSessions

Permissions

PermissionsConfig

PoolXYK

PriceTools

PswapDistribution

PswapDistributionConfig

RandomnessCollectiveFlip

Referrals

Rewards

RewardsConfig

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

Payload data to be signed when making signed transaction from off-chain workers, inside create_transaction function.

Staking

StakingConfig

System

SystemConfig

TechAccountId

TechAssetId

Technical

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tokens

TokensConfig

TradingPair

TradingPairConfig

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

VestedRewards

VestedRewardsConfig

Weight

Numeric range of a transaction weight.

XSTPool

XSTPoolConfig

XorFee