Enum framenode_runtime::FilterMode

pub enum FilterMode {
    Disabled,
    ForbidSelected,
    AllowSelected,
}

Variants

Disabled

Filter is disabled, all items regardless of filter are included.

ForbidSelected

Only selected items are filtered out, rest will be included.

AllowSelected

Only selected items will be included, rest are filtered out.