Trait framenode_runtime::LiquiditySource

pub trait LiquiditySource<TargetId, AccountId, AssetId, Amount, Error> {
    fn can_exchange(
        target_id: &TargetId, 
        input_asset_id: &AssetId, 
        output_asset_id: &AssetId
    ) -> bool;
    fn quote(
        target_id: &TargetId, 
        input_asset_id: &AssetId, 
        output_asset_id: &AssetId, 
        amount: QuoteAmount<Amount>, 
        deduce_fee: bool
    ) -> Result<SwapOutcome<Amount>, DispatchError>;
    fn exchange(
        sender: &AccountId, 
        receiver: &AccountId, 
        target_id: &TargetId, 
        input_asset_id: &AssetId, 
        output_asset_id: &AssetId, 
        swap_amount: SwapAmount<Amount>
    ) -> Result<SwapOutcome<Amount>, DispatchError>;
    fn check_rewards(
        target_id: &TargetId, 
        input_asset_id: &AssetId, 
        output_asset_id: &AssetId, 
        input_amount: Amount, 
        output_amount: Amount
    ) -> Result<Vec<(Amount, AssetId, RewardReason), Global>, DispatchError>;
    fn quote_without_impact(
        target_id: &TargetId, 
        input_asset_id: &AssetId, 
        output_asset_id: &AssetId, 
        amount: QuoteAmount<Amount>, 
        deduce_fee: bool
    ) -> Result<SwapOutcome<Amount>, DispatchError>;
}

Indicates that particular object can be used to perform exchanges.

Required Methods

fn can_exchange(
target_id: &TargetId,
input_asset_id: &AssetId,
output_asset_id: &AssetId
) -> bool

Check if liquidity source provides an exchange from given input asset to output asset.

fn quote(
target_id: &TargetId,
input_asset_id: &AssetId,
output_asset_id: &AssetId,
amount: QuoteAmount<Amount>,
deduce_fee: bool
) -> Result<SwapOutcome<Amount>, DispatchError>

Get spot price of tokens based on desired amount.

fn exchange(
sender: &AccountId,
receiver: &AccountId,
target_id: &TargetId,
input_asset_id: &AssetId,
output_asset_id: &AssetId,
swap_amount: SwapAmount<Amount>
) -> Result<SwapOutcome<Amount>, DispatchError>

Perform exchange based on desired amount.

fn check_rewards(
target_id: &TargetId,
input_asset_id: &AssetId,
output_asset_id: &AssetId,
input_amount: Amount,
output_amount: Amount
) -> Result<Vec<(Amount, AssetId, RewardReason), Global>, DispatchError>

Get rewards that are given for perfoming given exchange.

fn quote_without_impact(
target_id: &TargetId,
input_asset_id: &AssetId,
output_asset_id: &AssetId,
amount: QuoteAmount<Amount>,
deduce_fee: bool
) -> Result<SwapOutcome<Amount>, DispatchError>

Get spot price of tokens based on desired amount, ignoring non-linearity of underlying liquidity source.