Trait framenode_runtime::MultiplierUpdate

pub trait MultiplierUpdate: Convert<FixedU128, FixedU128> {
    // Required methods
    fn min() -> FixedU128;
    fn max() -> FixedU128;
    fn target() -> Perquintill;
    fn variability() -> FixedU128;
}

Something that can convert the current multiplier to the next one.

Required Methods

fn min() -> FixedU128

Minimum multiplier. Any outcome of the convert function should be at least this.

fn max() -> FixedU128

Maximum multiplier. Any outcome of the convert function should be less or equal this.

fn target() -> Perquintill

Target block saturation level

fn variability() -> FixedU128

Variability factor