Type Definition framenode_runtime::AllPalletsReversedWithSystemFirst

pub type AllPalletsReversedWithSystemFirst = (System, Beefy, SubstrateBridgeApp, MultisigVerifier, BridgeDataSigner, ParachainBridgeApp, SubstrateDispatch, SubstrateBridgeOutboundChannel, SubstrateBridgeInboundChannel, BridgeProxy, LeafProvider, OrderBook, Preimage, HermesGovernancePlatform, OracleProxy, Band, ElectionProviderMultiPhase, BagsList, DemeterFarmingPlatform, CeresLaunchpad, CeresGovernancePlatform, CeresTokenLocker, CeresLiquidityLocker, CeresStaking, PriceTools, XSTPool, Farming, Identity, VestedRewards, ElectionsPhragmen, TechnicalMembership, IrohaMigration, Scheduler, Multisig, PswapDistribution, EthBridge, DEXAPI, Democracy, TechnicalCommittee, Council, LiquidityProxy, PoolXYK, Technical, MulticollateralBondingCurvePool, DEXManager, Assets, TradingPair, Currencies, Tokens, ImOnline, Grandpa, Session, Historical, Offences, Staking, Authorship, Utility, BridgeMultisig, XorFee, Rewards, Referrals, Permissions, TransactionPayment, RandomnessCollectiveFlip, Balances, Timestamp, Babe);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.