Crate framenode_runtime

Re-exports

pub use assets;

pub use dex_api;

pub use eth_bridge;

pub use frame_system;

pub use kensetsu;

pub use liquidity_proxy;

pub use multicollateral_bonding_curve_pool;

pub use order_book;

pub use trading_pair;

pub use xst;

pub use oracle_types::*;

Modules

api

constants

Constant values used within the runtime.

migrations

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

oracle_types

Types used by oracle related pallets

weights

Macros

balance

construct_runtime

Construct a runtime, with the given name and the given pallets.

fixed

parameter_types

Create new implementations of the Get trait.

Structs

AlignmentSchedulerMaxWeight

ApolloOffchainWorkerTxLongevity

ApolloOffchainWorkerTxPriority

AssetName

AssetSymbol

BagThresholds

BalanceWrapper

Wrapper type which extends Balance serialization, used for json in RPC’s.

BandMaxRelaySymbols

BaseCallFilter

BasicDeposit

BetterUnsignedThreshold

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

BlockLength

BlockWeights

Block weights base values and limits.

BondingDuration

BridgeMaxGasPerMessage

BridgeMaxMessagePayloadSize

BridgeMaxMessagesPerCommit

BridgeMaxPeers

BridgeMaxTotalGasLimit

BuyBackTBCDPercent

CeresAssetId

CeresPerDay

ConstantFeeMultiplier

ContentSource

CouncilCollectiveMaxMembers

CouncilCollectiveMaxProposals

CouncilCollectiveMotionDuration

CreationFee

CrowdloanTag

DEXIdValue

DataSignerLongevity

DataSignerPriority

Decimals

DemeterAssetId

DemocracyCooloffPeriod

DemocracyEnactmentPeriod

DemocracyFastTrackVotingPeriod

DemocracyInstantAllowed

DemocracyLaunchPeriod

DemocracyMaxBlacklisted

DemocracyMaxDeposits

DemocracyMaxProposals

DemocracyMaxVotes

DemocracyMinimumDeposit

DemocracyPreimageByteDeposit

DemocracyVotingPeriod

DepositBase

DepositFactor

DescriptionLimit

DisabledValidatorsThreshold

ElectionBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionLookahead

ElectionsCandidacyBond

ElectionsDesiredMembers

13 members initially, to be increased to 23 eventually.

ElectionsDesiredRunnersUp

ElectionsMaxCandidates

ElectionsMaxVoters

ElectionsModuleId

ElectionsTermDuration

ElectionsVotingBondBase

ElectionsVotingBondFactor

EpochDuration

ExistentialDeposit

ExistentialDeposits

ExpectedBlockTime

ExpirationsSchedulerMaxWeight

FarmingRewardDoublingAssets

FeeCurrency

FieldDeposit

GenericTimepointProvider

GenesisConfig

GetADARAccountId

GetBandRateStaleBlockPeriod

GetBandRateStalePeriod

GetBaseAssetId

GetBurnUpdateFrequency

GetBuyBackAccountId

GetBuyBackAssetId

GetBuyBackDexId

GetBuyBackPercentage

GetBuyBackSupplyAssets

GetChameleonPool

GetChameleonPoolBaseAssetId

GetDaiAssetId

GetDefaultSubscriptionFrequency

GetEthAssetId

GetEthNetworkId

GetFarmingRewardsAccountId

GetFarmingRewardsTechAccountId

GetFee

GetInternalSlippageTolerancePercent

GetKarmaIncentiveRemintPercent

GetKenIncentiveRemintPercent

GetLiquidityProxyAccountId

GetLiquidityProxyTechAccountId

GetMarketMakerRewardsAccountId

GetMarketMakerRewardsTechAccountId

GetMbcPoolFreeReservesAccountId

GetMbcPoolFreeReservesTechAccountId

GetMbcPoolRewardsAccountId

GetMbcPoolRewardsTechAccountId

GetMbcReservesAccountId

GetMbcReservesTechAccountId

GetNumSamples

GetParliamentAccountId

GetPswapAssetId

GetPswapDistributionAccountId

GetPswapDistributionTechAccountId

GetReferenceAssetId

GetReferenceDexId

GetSyntheticBaseAssetId

GetSyntheticBaseBuySellLimit

GetTBCBuyBackTBCDPercent

GetTbcIrreducibleReservePercent

GetTbcdAssetId

GetTotalBalance

GetTradingPairRestrictedFlag

GetTreasuryAccountId

GetTreasuryTechAccountId

GetTrustlessBridgeAccountId

GetTrustlessBridgeFeesAccountId

GetTrustlessBridgeFeesTechAccountId

GetTrustlessBridgeTechAccountId

GetValAssetId

GetXSTPoolPermissionedAccountId

GetXSTPoolPermissionedTechAccountId

GetXorAssetId

GetXorFeeAccountId

GetXorFeeTechAccountId

GetXstAssetId

GetXstPoolConversionAssetId

GetXykIrreducibleReservePercent

HermesAssetId

ImOnlineUnsignedPriority

KarmaAssetId

KenAssetId

KensetsuOffchainWorkerTxLongevity

KensetsuOffchainWorkerTxPriority

KensetsuTreasuryAccountId

KensetsuTreasuryTechAccountId

LiquiditySourceFilter

Generalized filtration mechanism for listing liquidity sources.

LiquiditySourceId

Identification of liquidity source.

MaxActiveValidators

Setup election pallet to support maximum winners upto 1200. This will mean Staking Pallet cannot have active validators higher than this count.

MaxAdditionalDataLengthSwapTransferBatch

MaxAdditionalDataLengthXorlessTransfer

MaxAdditionalFields

MaxAuthorities

MaxElectableTargets

… and all of the validators as electable targets. Whilst this is the case, we cannot and shall not increase the size of the validator intentions.

MaxElectingVoters

We take the top 12500 nominators as electing voters..

MaxIterations

MaxKeys

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxRegistrars

MaxScheduledPerBlock

MaxSetIdSessionEntries

MaxSignatories

MaxSubAccounts

MaximumCeresInStakingPool

MinSolutionScoreBump

MinimalStabilityFeeAccrue

MinimumPeriod

MultiVerifier

NoPreimagePostponement

NposCompactSolution24

NposSolutionPriority

OffchainRepeat

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

OffencesWeightSoftLimit

OffendingValidatorsThreshold

OnChainSeqPhragmen

OperationalFeeMultiplier

OptionsLimit

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageBaseDeposit

PreimageByteDeposit

PresetWeightInfo

ReferralsReservesAcc

ReferrerWeight

RemovePendingOutgoingRequestsAfter

RemoveTemporaryPeerAccountIds

ReportLongevity

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeOnPswapBurnedAggregator

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SchedulerMaxWeight

SessionDuration

SessionOffset

SessionPeriod

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxRefunds

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

StakingBenchmarkingConfig

A reasonable benchmarking config for staking pallet.

StringLimit

SubAccountDeposit

SwapOutcome

Amount of output tokens from either price request or actual exchange.

TechnicalCollectiveMaxMembers

TechnicalCollectiveMaxProposals

TechnicalCollectiveMotionDuration

ThisNetworkId

TitleLimit

TrackPendingIncomingRequestsAfter

TransactionByteFee

TransferFee

U128CurrencyToVote

An implementation of CurrencyToVote tailored for chain’s that have a balance type of u128.

UnsignedPhase

ValBurnedAggregator

ValRewardCurve

Version

Weight

XorBurnedWeight

XorIntoValBurnedWeight

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

DispatchClass

A generalized group of dispatch types.

FilterMode

LiquiditySourceType

Enumaration of all available liquidity sources.

MultiProof

OriginCaller

RuntimeCall

RuntimeEvent

StakerStatus

Indicates the initial status of the staker.

SwapAmount

Used to identify intention of caller either to transfer tokens based on exact input amount or exact output amount.

SwapVariant

Used to identify intention of caller to indicate desired input amount or desired output amount. Similar to SwapAmount, does not hold value in order to be used in external API.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

FARMING_PSWAP_PER_DAY

FARMING_REFRESH_FREQUENCY

FARMING_VESTING_COEFF

FARMING_VESTING_FREQUENCY

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

AssetInfoProvider

Trait to provide info about assets

BuildStorage

Complex storage builder stuff.

Contains

A trait for querying whether a type can be said to “contain” a value.

DexInfoProvider

Trait to provide DEXInfo

FromGenericPair

Abstract trait to get data type from generic pair name and data.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

LiquiditySource

Indicates that particular object can be used to perform exchanges.

MultiplierUpdate

Something that can convert the current multiplier to the next one.

OnPswapBurned

General trait for passing pswap amount burned information to required pallets.

OnUnbalanced

Handler for when some currency “account” decreased in balance for some reason.

OnValBurned

General trait for passing on the amount of burned VAL.

Randomness

A trait that is able to provide randomness.

ScheduleNamed

A type that can be used as a scheduler.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

SyntheticInfoProvider

TradingPairSourceManager

Functions

debug

Print out the debuggable type.

fixed_from_basis_points

Create fraction as Fixed from BasisPoints value.

native_version

The version infromation used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Amount

ApolloPlatform

AssetId

Assets

AssetsConfig

Authorship

Babe

BabeConfig

BagsList

Balance

BalancePrecision

Balances

BalancesConfig

Band

BasisPoints

Type definition representing financial basis points (1bp is 0.01%)

Beefy

BeefyConfig

BeefyId

Identity of a BEEFY authority using ECDSA as its crypto.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

BridgeDataSigner

BridgeMultisig

BridgeMultisigConfig

BridgeProxy

CeresGovernancePlatform

CeresLaunchpad

CeresLiquidityLocker

CeresStaking

CeresTokenLocker

CheckedExtrinsic

Extrinsic type that has already been checked.

Council

CouncilCollective

CouncilConfig

Currencies

DEXAPI

DEXAPIConfig

DEXId

Identification of DEX.

DEXManager

DEXManagerConfig

DemeterFarmingPlatform

Democracy

DemocracyConfig

DigestItem

Digest item type.

ElectionProviderMultiPhase

ElectionsPhragmen

ElectionsPhragmenConfig

EthBridge

EthBridgeConfig

Executive

Executive: handles dispatch to the various modules.

ExtendedAssets

Farming

Fixed

Basic type representing assets quantity.

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

HermesGovernancePlatform

Historical

Identity

ImOnline

ImOnlineConfig

ImOnlineId

An i’m online identifier using sr25519 as its crypto.

Index

Index of a transaction in the chain.

IrohaMigration

IrohaMigrationConfig

Kensetsu

KensetsuConfig

LeafProvider

LiquidityProxy

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Moment

MulticollateralBondingCurvePool

MulticollateralBondingCurvePoolConfig

Multiplier

Fee multiplier.

Multisig

MultisigVerifier

NetworkId

Offences

OnChainAccuracy

The accuracy type used for genesis election provider;

OracleProxy

OrderBook

ParachainBridgeApp

ParachainBridgeAppConfig

PeriodicSessions

Permissions

PermissionsConfig

PoolXYK

Preimage

PriceTools

PswapDistribution

PswapDistributionConfig

RandomnessCollectiveFlip

Referrals

Rewards

RewardsConfig

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

Payload data to be signed when making signed transaction from off-chain workers, inside create_transaction function.

Staking

StakingConfig

SubstrateBridgeApp

SubstrateBridgeInboundChannel

SubstrateBridgeOutboundChannel

SubstrateBridgeOutboundChannelConfig

SubstrateDispatch

System

SystemConfig

TechAccountId

TechAssetId

Technical

TechnicalCollective

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tokens

TokensConfig

TradingPair

TradingPairConfig

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

VestedRewards

XSTPool

XSTPoolConfig

XorFee